+++ src/openai/models.ts
@@ -10,12 +10,13 @@ export const modelData: OpenAI.Models.Model[] = Object.keys(ModelMapping).map((m
 
 export const models = async (req: Request) => {
   const apiParam = getToken(req.headers)
-  return await listModels(apiParam)
-
-  // return {
-  //   object: "list",
-  //   data: modelData,
-  // }
+  var gemini_models = await listModels(apiParam);
+  var model_list = gemini_models["data"] as Array<any>
+  const finala_model_list=model_list.concat(modelData);
+  return {
+    object: "list",
+    data: finala_model_list,
+  }
 }
 
 export const modelDetail = (model: string) => {